backend/requirements.txt
Django==4.1.5
djangorestframework
djoser==2.1.0
gunicorn==20.1.0
psycopg2-binary==2.8.6
